Send an 'Accept' header in the image request
Some urls [1] don't respond correctly if there's no 'Accept' header in
the request. This change adds an 'Accept: image/*' header to each
request for external images. I'm not 100% sure that this is the correct
and/or best thing to be doing, but it certainly fixes it for the
specific case of images on www.congreso.gob.gt.

1: e.g. this one -
http://www.congreso.gob.gt/manager/images/69DBC2D6-2EE6-3EA7-411F-07C91997560D.jpg