Convert the image if a format is specified
Currently, if you include an 'extension' in your request, the
Content-type of the response will match that format, no matter what
image format the source image was.

However, this doesn't actually reformat the image. It just changes the
Content-type on the response headers.

Reformatting the actually image provides more control, such as accessing
the nice image compression you get from the jpeg format.

This commit passes the 'extension' to the .stream() function.
The image is streamed back in the specified format.

If no 'extension' has been specified by the user, it falls back to the
Content-type of the source image.